Fixed issue related to Search/Undo displayed empty list data
