+++ app/src/main/java/com/todo/notesapp/utils/NotesUtils.kt
@@ -8,17 +8,6 @@ import androidx.fragment.app.Fragment
 
 object NotesUtils {
 
-    fun hideKeyboard(activity: Activity) {
-        val inputManager = activity.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
-
-        val currentFocusedView = activity.currentFocus
-        currentFocusedView.let {
-            inputManager.hideSoftInputFromWindow(
-                currentFocusedView?.windowToken, InputMethodManager.HIDE_NOT_ALWAYS
-            )
-        }
-    }
-
     @JvmStatic
     fun Fragment.hideKeyboard() {
         view?.let { activity?.hideKeyboard(it) }
